+++ src/pipecatcloud/cli/organizations.py
@@ -2,18 +2,20 @@ import aiohttp
 import questionary
 import typer
 from rich import box
-from rich.console import Console
-from rich.panel import Panel
 from rich.table import Table
 
 from pipecatcloud import PIPECAT_CLI_NAME
 from pipecatcloud._utils.async_utils import synchronizer
 from pipecatcloud._utils.auth_utils import requires_login
-from pipecatcloud._utils.console_utils import print_api_error
+from pipecatcloud._utils.console_utils import console, print_api_error
 from pipecatcloud._utils.http_utils import construct_api_url
-from pipecatcloud.cli import PANEL_TITLE_SUCCESS
-from pipecatcloud.config import _store_user_config, config, user_config_path
-from pipecatcloud.exception import AuthError
+from pipecatcloud.api import API
+from pipecatcloud.config import (
+    _store_user_config,
+    dashboard_host,
+    update_user_config,
+    user_config_path,
+)
 
 organization_cli = typer.Typer(
     name="organizations", help="User organizations.", no_args_is_help=True
@@ -22,30 +24,10 @@ keys_cli = typer.Typer(name="keys", help="API key management commands.", no_args
 organization_cli.add_typer(keys_cli)
 
 
-# ---- Organization Methods ----
+# ---- Methods
 
-async def _retrieve_organizations(ctx: typer.Context):
-    console = Console()
-    token = ctx.obj["token"]
-    org_list = []
-    with console.status("Fetching user organizations", spinner="dots"):
-        async with aiohttp.ClientSession() as session:
-            try:
-                async with session.get(
-                    f"{config.get('server_url')}{config.get('organization_path')}",
-                    headers={"Authorization": f"Bearer {token}"},
-                ) as resp:
-                    if resp.status == 401:
-                        raise AuthError()
-                    if resp.status == 200:
-                        data = await resp.json()
-                        org_list = data["organizations"]
-                    else:
-                        raise Exception(f"Failed to retrieve account organization: {resp.status}")
-            except Exception as e:
-                console.print(f"[red]Error: {e}[/red]")
-
-    return org_list
+def _set_key_as_default(org: str, key_name: str, key_value: str):
+    update_user_config({"default_public_key": key_value, "default_public_key_name": key_name}, org)
 
 
 async def _get_api_tokens(org_id: str, token: str):
@@ -64,61 +46,94 @@ async def _get_api_tokens(org_id: str, token: str):
         return data
 
 
-# ---- Organization Commands ----
+# ---- Commands
 @organization_cli.command(name="select", help="Select an organization to use.")
 @synchronizer.create_blocking
 @requires_login
-async def select(ctx: typer.Context):
-    console = Console()
+async def select(
+    ctx: typer.Context,
+    organization: str = typer.Option(
+        None,
+        "--organization",
+        "-o"
+    )
+):
     current_org = ctx.obj["org"]
-    org_list = await _retrieve_organizations(ctx)
 
-    value = await questionary.select(
-        "Select active organization",
-        choices=[{"name": f"{org['verboseName']} ({org['name']})", "value": (org["name"], org["verboseName"]), "checked": org["name"] == current_org} for org in org_list],
-    ).ask_async()
+    with console.status("[dim]Retrieve user namespace / organization data...[/dim]", spinner="dots"):
+        org_list, error = await API.organizations()
 
-    if not value:
-        return
+        if error:
+            typer.Exit()
 
-    _store_user_config(ctx.obj["token"], value[0])
+    try:
+        selected_org = None, None
+        if not organization:
+            # Prompt user to select organization
+            value = await questionary.select(
+                "Select default namespace / organization",
+                choices=[{"name": f"{org['verboseName']} ({org['name']})", "value": (org["name"], org["verboseName"]), "checked": org["name"] == current_org} for org in org_list],
+            ).ask_async()
 
-    console.print(Panel(
-        f"Current organization set to [bold green]{value[1]} [dim]({value[0]})[/dim][/bold green]\n"
-        f"[dim]Account updated in {user_config_path}[/dim]",
-        title="[green]Organization updated[/green]",
-        title_align="left",
-        border_style="green",
-    ))
+            if not value:
+                return typer.Exit(1)
+
+            selected_org = value[0], value[1]
+
+        else:
+            # Attempt to match passed org with results
+            match = None
+            for o in org_list:
+                if o["name"] == organization:
+                    match = o
+            if not match:
+                console.error(
+                    f"Unable to find namespace [bold]'{organization}'[/bold] in user's available organizations"
+                )
+                return typer.Exit(1)
+            selected_org = match["name"], match["verboseName"]
+
+        _store_user_config(ctx.obj["token"], selected_org[0])
+
+        console.success(
+            f"Current organization set to [bold green]{selected_org[1]} [dim]({selected_org[0]})[/dim][/bold green]\n"
+            f"[dim]Default namespace updated in {user_config_path}[/dim]")
+    except Exception:
+        console.error("Unable to update user credentials. Please contact support.")
 
 
 @organization_cli.command(name="list", help="List organizations user is a member of.")
 @synchronizer.create_blocking
 @requires_login
 async def list(ctx: typer.Context):
-    console = Console()
     current_org = ctx.obj["org"]
-    org_list = await _retrieve_organizations(ctx)
+    with console.status("[dim]Retrieve user namespace / organization data...[/dim]", spinner="dots"):
+        org_list, error = await API.organizations()
 
-    if len(org_list) == 0:
-        console.print("[red]No organizations found[/red]")
-        return
-    else:
-        console.print(f"[green]Found {len(org_list)} organizations[/green]")
+        if error:
+            return typer.Exit()
+
+    if not org_list or not len(org_list):
+        console.error(
+            "No namespaces associated with user account. Please complete onboarding via the dashboard.",
+            subtitle=dashboard_host)
+        return typer.Exit(1)
 
     table = Table(
         border_style="dim",
+        box=box.SIMPLE,
         show_edge=True,
         show_lines=False)
     table.add_column("Organization", style="white")
     table.add_column("Name", style="white")
     for org in org_list:
         if org["name"] == current_org:
-            table.add_row(org["verboseName"], f"[cyan bold]{org['name']} (active)[/cyan bold]")
+            table.add_row(f"[cyan bold]{org['verboseName']}[/cyan bold]",
+                          f"[cyan bold]{org['name']} (active)[/cyan bold]")
         else:
             table.add_row(org["verboseName"], org["name"])
 
-    console.print(table)
+    console.success(table, title_extra=f"{len(org_list)} results")
 
 
 # ---- API Token Commands ----
@@ -131,24 +146,106 @@ async def keys(
     organization: str = typer.Option(
         None,
         "--organization",
-        "--org",
-        help="Organization to get tokens for",
+        "-o",
+        help="Organization to list API keys for",
     ),
 ):
-    console = Console()
-    token = ctx.obj["token"]
     org = organization or ctx.obj["org"]
 
-    with console.status(f"Fetching API keys for organization: [bold]'{org}'[/bold]", spinner="dots"):
-        data = await _get_api_tokens(org, token)
+    try:
+        with console.status(f"[dim]Fetching API keys for organization: [bold]'{org}'[/bold][/dim]", spinner="dots"):
+            data = await API.api_keys(org)
 
-    if len(data["public"]) == 0:
-        console.print(
-            f"[bold]No API keys found.[/bold]\n"
-            f"[dim]Create a new API key with the "
-            f"[bold]{PIPECAT_CLI_NAME} organizations keys create[/bold] command.[/dim]"
-        )
-        return
+            if len(data["public"]) == 0:
+                console.error(
+                    f"[bold]No API keys found.[/bold]\n"
+                    f"[dim]Create a new API key with the "
+                    f"[bold]{PIPECAT_CLI_NAME} organizations keys create[/bold] command.[/dim]"
+                )
+                return typer.Exit(1)
+
+            table = Table(
+                show_header=True,
+                show_lines=True,
+                border_style="dim",
+                box=box.SIMPLE,
+            )
+            table.add_column("Name")
+            table.add_column("Key")
+            table.add_column("Created At")
+            table.add_column("Status")
+
+            for key in data["public"]:
+                table.add_row(
+                    key["metadata"]["name"],
+                    key["key"],
+                    key["createdAt"],
+                    "Revoked" if key["revoked"] else "Active",
+                    style="red" if key["revoked"] else None,
+                )
+
+            console.success(table, title_extra=f"API keys for organization: {org}")
+    except Exception:
+        console.error("Failed to fetch API keys. Please contact support")
+
+
+@keys_cli.command(name="create", help="Create an API key for an organization.")
+@synchronizer.create_blocking
+@requires_login
+async def create_key(
+    ctx: typer.Context,
+    api_key_name: str = typer.Option(
+        None,
+        "--name",
+        "-n",
+        help="Human readable name for new API key",
+    ),
+    organization: str = typer.Option(
+        None,
+        "--organization",
+        "-o",
+        help="Organization to create API key for",
+    ),
+    default: bool = typer.Option(
+        False,
+        "--default",
+        "-d",
+        help="Set the newly created key as the active / default key in local config"
+    )
+):
+    org = organization or ctx.obj["org"]
+
+    if not api_key_name:
+        api_key_name = await questionary.text(
+            "Enter human readable name for API key e.g. 'Pipecat Key'"
+        ).ask_async()
+
+    if not api_key_name or api_key_name == "":
+        console.error("You must enter a name for the API key")
+        return typer.Exit(1)
+
+    error_code = None
+    data = None
+    try:
+        with console.status(f"[dim]Creating API key with name: [bold]'{api_key_name}'[/bold][/dim]", spinner="dots"):
+            data = await API.api_key_create(api_key_name, org)
+    except Exception:
+        console.api_error(error_code, title="Error creating API key")
+        return typer.Exit(1)
+
+    if not data or 'key' not in data:
+        console.error("Invalid response from server. Please contact support.")
+        return typer.Exit(1)
+
+    # Determine as to whether we should make this key the active default
+    make_active = default
+    if not default:
+        make_active = await questionary.confirm("Would you like to make this key the default key in your local configuration?", default=False).ask_async()
+
+    if make_active:
+        _set_key_as_default(org, api_key_name, data['key'])
+    else:
+        console.print("[dim]Bypassing using key as default")
 
     table = Table(
         show_header=True,
@@ -158,23 +255,15 @@ async def keys(
     )
     table.add_column("Name")
     table.add_column("Key")
-    table.add_column("Created At")
-    table.add_column("Status")
+    table.add_column("Organization")
 
-    for key in data["public"]:
-        table.add_row(
-            key["metadata"]["name"],
-            key["key"],
-            key["createdAt"],
-            "Revoked" if key["revoked"] else "Active",
-            style="red" if key["revoked"] else None,
-        )
+    table.add_row(
+        api_key_name,
+        data['key'],
+        org,
+    )
 
-    console.print(Panel(
-        table,
-        title=f"[bold]API keys for organization: {org}[/bold]",
-        title_align="left",
-    ))
+    console.success(table)
 
 
 @keys_cli.command(name="delete", help="Delete an API key for an organization.")
@@ -230,80 +319,6 @@ async def delete_key(
         typer.Exit(1)
 
     console.print(f"[green]API key with ID: [bold]'{key_id}'[/bold] deleted successfully.[/green]")
-
-
-@keys_cli.command(name="create", help="Create an API key for an organization.")
-@synchronizer.create_blocking
-@requires_login
-async def create_key(
-    ctx: typer.Context,
-    organization: str = typer.Option(
-        None,
-        "--organization",
-        "--org",
-        help="Organization to get tokens for",
-    ),
-):
-    console = Console()
-    token = ctx.obj["token"]
-    org = organization or ctx.obj["org"]
-
-    api_key_name = await questionary.text(
-        "Enter human readable name for API key e.g. 'Pipecat Key'"
-    ).ask_async()
-
-    if not api_key_name:
-        typer.Exit(1)
-
-    error_code = None
-    data = None
-    try:
-        with console.status(f"Creating API key with name: [bold]'{api_key_name}'[/bold]", spinner="dots"):
-            async with aiohttp.ClientSession() as session:
-                response = await session.post(
-                    f"{construct_api_url('api_keys_path').format(org=org)}",
-                    headers={"Authorization": f"Bearer {token}"},
-                    json={"name": api_key_name, "type": "public"},
-                )
-                if response.status != 200:
-                    error_code = str(response.status)
-                    response.raise_for_status()
-                data = await response.json()
-    except Exception:
-        print_api_error(error_code, title="Error creating API key")
-        typer.Exit(1)
-        return
-
-    if not data or 'key' not in data:
-        console.print("[red]Error: Invalid response from server[/red]")
-        typer.Exit(1)
-        return
-
-    table = Table(
-        show_header=True,
-        show_lines=True,
-        border_style="dim",
-        box=box.SIMPLE,
-    )
-    table.add_column("Name")
-    table.add_column("Key")
-    table.add_column("Organization")
-
-    table.add_row(
-        api_key_name,
-        data['key'],
-        org,
-    )
-
-    console.print(
-        Panel(
-            table,
-            title=f"[green]{PANEL_TITLE_SUCCESS}[/green]",
-            title_align="left",
-            border_style="green",
-            subtitle=f"Use the key by default by running [bold]{PIPECAT_CLI_NAME} organizations keys use[/bold]",
-            subtitle_align="left",
-        ))
 
 
 @keys_cli.command(name="use", help="Set default API key for an organization in local config.")